Basic functionality tested
Plugged a hodgepodge cable in from one USB to another and verified that
this code can read data coming in. Also changed serial output from
label to text box so that it can scroll and not run off the page for
lots of output. Got log file creation working but not yet writing to
it.